Replace `RecordIterator` with `ParserIterator`
`ParserIterator` is identical to its predecessor except that it includes
all of the parse methods (e.g. `record`, `row`) that previously belonged
to `Parser`.
By moving these methods to the iterator instead of the `Parser` object,
we get to:
A. Simplify the interface, since `record` etc. no longer have to be
   generic.
B. Don't have to mess around with cloned `Parser`s when creating
   iterators via `parse`.
   With this new setup, `Parser` can be treated as immutable once its
   been built. This lets us more easily reuse the same `Parser` object
   when parsing different files.