Refactor `ParserIterator` to use strategies
Problem:
Configuration is handled by various conditionals interspersed with the
parsing logic. This makes the code messy, hard to read, and hard to
change.

Solution:
`ParserIterator` now uses something like the 'strategy pattern' to
handle configuration. When a 'ParserIterator' is created, the
configuration options are used to set different callbacks that are used
by in the parsing logic. This helps with the above as each strategy
callback is self-contained. Adding new configuration just involves
creating a new strategy and assigning it during `new`.