Fix bug in `read_line` and more perf improvements
Bug:
The `CsvReader::read_line` method is supposed to return `None` if
nothing was read. It (incorrectly) checks if it should do this by
checking if the length of the buffer it's reading to is 0. This is
wrong, as it means that an initial non-empty buffer will cause the
method to never return `None`. Instead, it should keep track of the
bytes its read and return `None` if that number is 0.

Performance:
I've substantially sped up `quote` and `text` by reducing the number
of times `buf.len_sans_newline()` is called. With these changes, our
parser now performs comparably to Sushi's.