Split `Records` into `Parser` and `Records`
Having a separate `Parser` struct makes it possible for users to
interact with the parser outside of iteration. For example, we plan to
add a `header` method so that users can get a CSV file's column names.
Iterating by records is now performed by calling the `records` method on
a `Parser`.