Add `headers` method to `Parser`
The `headers` method lets users request a CSV file's headers. Headers
correspond to the fields of the first record read by `Parser`. Note
that `headers` can be called at any time and does *not* require
`detect_headers` to be on. As such, `Parser`s must always track the
first record that it reads.