Improve `Record` interface and safety
Interface:
`Record` no longer exposes its `field_bounds` member directly. Instead,
it now has `num_fields` and `set_num_fields` methods.

Safety:
`Record` now uses a `String` for its buffer instead of a `Vec<u8>`.
Previously, it was assumed that the bytes passed to `Record::new` were
valid UTF-8. By using `String`, we have compile-time guarantees that
this is the case.
As a part of this change, `RecordBuffer::take_inner` has been renamed to
`RecordBuffer::take_inner_as_string`. This method performs an unchecked
conversion of its inner byte vector to a `String`. This is ok because
`RecordBuffer` validates that this vector contains only UTF-8 whenever
it is read to.