Simplify return values used in parsing
While parsing a record, we want to be able to indicate 3 things:
1. Parsing resulted in a record.
2. Reader is empty so parsing should stop.
3. Parsing resulted in an error.

Currently, we do this by always returning an Option<Result<T>> from our
functions and translating Some(Ok(T)), Some(Err(T)), and None to the
above states. This format was chosen because it mimics the return value
of iterators that can error.

The problem with this setup is that it makes it more cumbersome to
deal with intermediate values.

Fortunately, it turns out that we can mostly substitute
Option<Result<T>> with Result<T>, as only records (not fields, etc) can
be None.